requirements.txt
python-telegram-bot==22.6
yt-dlp>=2024.1.0
python-dotenv==1.0.0
